+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Cliprompt Changelog
 v0.0.2 - wip
 -----------
 
+- fix case when numbers are used in choices
 - added some test coverage for further evolution easiness
 
 v0.0.1 - 2014-05-19